AI_Service/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
google-genai
json_repair


